+++ tests/test_function/test_parse_scope.py
@@ -8,6 +8,19 @@ from pycep import BicepParser
 EXAMPLES_DIR = Path(__file__).parent / "examples/scope"
 
 
+def test_parse_management_group() -> None:
+    # given
+    sub_dir_path = EXAMPLES_DIR / "management_group"
+    file_path = sub_dir_path / "main.bicep"
+    expected_result = json.loads((sub_dir_path / "result.json").read_text())
+
+    # when
+    result = BicepParser(file_path=file_path).json()
+
+    # then
+    assert_that(result).is_equal_to(expected_result)
+
+
 def test_parse_resource_group() -> None:
     # given
     sub_dir_path = EXAMPLES_DIR / "resource_group"