+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Support `min` function
 - Support `int` function
 - Support `skip` function
+- Support `managementGroup` function
 
 ## 0.2.0 - 2022-02-13